00_RESTful URL setting

+++ boards/views.py
@@ -1,7 +1,18 @@
 from django.shortcuts import render, redirect
 from .models import Board
+from pprint import pprint
+
 # Create your views here.
 def index(request):
+    # pprint(request)
+    # pprint(type(request))
+    # pprint(dir(request))
+    # pprint(request.scheme)
+    # pprint(request.get_host())
+    # pprint(request.get_full_path())
+    # pprint(request.build_absolute_uri())
+    # pprint(request.META)
+    
     boards = Board.objects.order_by('-pk')
     context = {
         'boards': boards,
@@ -9,15 +20,15 @@ def index(request):
     return render(request, 'boards/index.html', context)
 
 def new(request):
-    return render(request, 'boards/new.html')
-
-def create(request):
-    title = request.POST.get('title')
-    content = request.POST.get('content')
-    
-    board = Board(title=title, content=content)
-    board.save()
-    return redirect('boards:detail', board.pk)
+    if request.method == 'POST':
+        title = request.POST.get('title')
+        content = request.POST.get('content')
+        
+        board = Board(title=title, content=content)
+        board.save()
+        return redirect('boards:detail', board.pk)
+    else:
+        return render(request, 'boards/new.html')
     
 def detail(request, pk):
     board = Board.objects.get(pk=pk)
@@ -28,19 +39,23 @@ def detail(request, pk):
 
 def delete(request, pk):
     board = Board.objects.get(pk=pk)
-    board.delete()
-    return redirect('boards:index')
+    if request.method == 'POST':
+        board.delete()
+        return redirect('boards:index')
+    else:
+        return redirect('boards:detail', board.pk)
     
 def edit(request, pk):
     board = Board.objects.get(pk=pk)
-    context = {
-        'board': board
-    }
-    return render(request, 'boards/edit.html', context)
-
-def update(request, pk):
-    board = Board.objects.get(pk=pk)
-    board.title = request.POST.get('title')
-    board.content = request.POST.get('content')
-    board.save()
-    return redirect('boards:detail', board.pk)
+    if request.method == 'POST':
+        # update
+        board.title = request.POST.get('title')
+        board.content = request.POST.get('content')
+        board.save()
+        return redirect('boards:detail', board.pk)
+    else:
+        # edit
+        context = {
+            'board': board
+        }
+        return render(request, 'boards/edit.html', context)
